pyproject.toml
[tool.poetry]
name = "pr-tracking"
version = "0.1.0"
description = ""
authors = ["Dario <dariiio.beltran@gmail.com>"]
readme = "README.md"
packages = [{include = "pr_tracking"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
django-rest-framework = "^0.1.0"
pillow = "^9.4.0"
psycopg2-binary = "^2.9.6"
crispy-bootstrap5 = "^0.7"
django-crispy-forms = "^2.0"
pre-commit = "^3.6.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
dj-database-url = "^2.2.0"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
